Extract Navbar Components and Refactor
