+++ src/pages/Journal/Navbar/Brand.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import styled from "styled-components";
+
+import { onPrimary } from "../../../common/styles";
+import { H1 } from "../../../components/Typeography";
+
+export const Brand = styled(H1)`
+  grid-area: left;
+  color: ${onPrimary};
+`;
