+++ src/pages/Journal/Navbar/HoverIcon.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import styled from "styled-components";
+
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+
+import { onPrimary, onPrimaryHover } from "../../../common/styles";
+
+export const HoverIcon = styled(FontAwesomeIcon)`
+  color: ${onPrimary};
+  &:hover {
+    color: ${onPrimaryHover};
+    cursor: pointer;
+  }
+`;
