+++ tslint.json
@@ -15,6 +15,7 @@
     "interface-name": true,
     "array-type": false,
     "jsx-no-multiline-js": false,
+    "jsx-boolean-value": false,
     "no-unused-variable": true
   },
   "rulesDirectory": []
